Refactored dataProvider
