--- pharmd-app/src/dataProvider.js
@@ -1,145 +1,0 @@
-import { stringify } from "query-string";
-import { fetchUtils } from "react-admin";
-import  { FAKE_API } from "./config/backendRoutes";
-
-// import real BACKEND_URL to use pharmD backend data
-const BACKEND_URL = FAKE_API;
-
-// Add authorization token to each request
-const httpClient = (url, options = {}) => {
-  if (!options.headers) {
-    options.headers = new Headers({ Accept: "application/json" });
-  }
-  const token = localStorage.getItem("token");
-  options.headers.set("Authorization", `Bearer ${token}`);
-  return fetchUtils.fetchJson(url, options);
-};
-
-export default {
-  getList: (resource, params) => {
-    const { page, perPage } = params.pagination;
-    const { field, order } = params.sort;
-    console.log("PARAMS", params);
-    const query = {
-      ...fetchUtils.flattenObject(params.filter),
-      _sort: field,
-      _order: order,
-      _start: (page - 1) * perPage,
-      _end: page * perPage
-    };
-    const url = `${BACKEND_URL}/${resource}?${stringify(query)}`;
-    console.log("URL", url);
-    return httpClient(url).then(({ headers, json, status }) => {
-      console.log("HAS AUTH", status);
-      if (!headers.has("x-total-count")) {
-        return Promise.reject("The X-Total-Count header is missing in the HTTP Response. The jsonServer Data Provider expects responses for lists of resources to contain this header with the total number of results to build the pagination. If you are using CORS, did you declare X-Total-Count in the Access-Control-Expose-Headers header?")
-      }
-      return {
-        data: json,
-        total: parseInt(
-          headers
-            .get("x-total-count")
-            .split("/")
-            .pop(),
-          10
-        )
-      };
-    });
-  },
-
-  getOne: (resource, params) =>
-    httpClient(`${BACKEND_URL}/${resource}/${params.id}`).then(({ json }) => ({
-      data: json
-    })),
-
-  getMany: (resource, params) => {
-    const query = {
-      id: params.ids
-    };
-    const url = `${BACKEND_URL}/${resource}?${stringify(query)}`;
-    return httpClient(url).then(({ json }) => ({ data: json }));
-  },
-
-  getManyReference: (resource, params) => {
-    const { page, perPage } = params.pagination;
-    const { field, order } = params.sort;
-    const query = {
-      ...fetchUtils.flattenObject(params.filter),
-      [params.target]: params.id,
-      _sort: field,
-      _order: order,
-      _start: (page - 1) * perPage,
-      _end: page * perPage
-    };
-    const url = `${BACKEND_URL}/${resource}?${stringify(query)}`;
-
-    return httpClient(url).then(({ headers, json }) => {
-      if (!headers.has("x-total-count")) {
-        return Promise.reject("The X-Total-Count header is missing in the HTTP Response. The jsonServer Data Provider expects responses for lists of resources to contain this header with the total number of results to build the pagination. If you are using CORS, did you declare X-Total-Count in the Access-Control-Expose-Headers header?")
-      }
-      return {
-        data: json,
-        total: parseInt(
-          headers
-            .get("x-total-count")
-            .split("/")
-            .pop(),
-          10
-        )
-      };
-    });
-  },
-
-  update: (resource, params) =>
-    httpClient(`${BACKEND_URL}/${resource}/${params.id}`, {
-      method: "PUT",
-      body: JSON.stringify(params.data)
-    }).then(({ json }) => ({ data: json })),
-
-  // json-server doesn't handle filters on UPDATE route, so we fallback to calling UPDATE n times instead
-  updateMany: (resource, params) =>
-    Promise.all(
-      params.ids.map(id =>
-        httpClient(`${BACKEND_URL}/${resource}/${id}`, {
-          method: "PUT",
-          body: JSON.stringify(params.data)
-        })
-      )
-    ).then(responses => ({ data: responses.map(({ json }) => json.id) })),
-
-  create: (resource, params) =>
-    httpClient(`${BACKEND_URL}/${resource}`, {
-      method: "POST",
-      body: JSON.stringify(params.data)
-    }).then(({ json }) => ({
-      data: { ...params.data, id: json.id }
-    })),
-
-  edit: (resource, params) => {
-    console.log("editing...", resource, params)
-    return httpClient(`${BACKEND_URL}/${resource}`, {
-      method: "PUT",
-      body: JSON.stringify(params.data)
-    }).then(({json}) => {
-      console.log(json)
-      return ({
-        data: {...params.data, id: json.id}
-      });
-    });
-  },
-
-  delete: (resource, params) =>
-    httpClient(`${BACKEND_URL}/${resource}/${params.id}`, {
-      method: "DELETE"
-    }).then(({ json }) => ({ data: json })),
-
-  // json-server doesn't handle filters on DELETE route, so we fallback to calling DELETE n times instead
-  deleteMany: (resource, params) =>
-    Promise.all(
-      params.ids.map(id =>
-        httpClient(`${BACKEND_URL}/${resource}/${id}`, {
-          method: "DELETE"
-        })
-      )
-    ).then(responses => ({ data: responses.map(({ json }) => json.id) }))
-};
