+++ pharmd-app/src/services/index.js
@@ -1,8 +1,10 @@
 import FileUploadService from "./FileUploadService";
 import AuthProvider from './Auth';
+import DataProvider from './Data';
 
 // This should be kept up-to-date as new service files are added
 export {
   FileUploadService,
-  AuthProvider
+  AuthProvider,
+  DataProvider
 };
